New class for PEEM voltages
A new class AcquisitionParametersVoltages (implements Serializable) is
created to storen the PEEM voltages of the AcquisitionParameters. There
are two ways to access the values:
1. The old fashioned way "ap.voltages.mcpU", however, now voltages has
be inserted
2. Getters and Setters are used to shorten the access "ap.getMcpU()"
For now only the voltages are stored in an own class, because they are
supposed to be serialized independently from the rest of the
AcquisitionParamters. In order to keep the code more consistent, the
currents and other fields in AcquisitionParamters should maybe also be
accessed through getters and setters