New saving system using serialization
The newly added class AcquisitionParametersVoltages implements the
Serializable interface. The load and save buttons in the
PEEMStateController have now been changed to use Java's intern
(De-/)Serialization to save the presets (files end with .pst).
The AcquisitionParameters that go with the saved PEEM images stay .txt
files because it's necessary to actually read them in the text editor