Saving the optimisation series
In order to save the optimisation series, the images taken are saved in
a list and titled with the property to be optimised and the current
value of the respective property. (The actual saving directory will
probably be changed in the future).
After the series is finished (and hasn't been cancelled), the current
AcquisitionParameters are recorded once. Since only the optimised value
changes significantly, the AcquisitionParameters for the image set are
artifically constructed instead of obtaining them for every single image
taken (as a result the other AcquisitionParameters are considered
constant).
The user is asked once (after the optimisation series is finished) to
enter the general data (excitation) and notes (similar to the single
image acquisition). If the user denies to enter any data or doesn't want
to save the images, the image set isn't saved.