Changing the saving directories for optimisation series
The optimisation series are saved in a pattern similar to the (single
shot) quickacquisition. They are marked by an additional "OptSeries" and
the optimised PeemProperty in the directory name.
The file names also contain the current value of the property. The
numbering of the optimisation series images doesn't work as well as the
quickaquisition, since they are ordered (in the saving folder) according
to the property value. However, the numbering of the saved images and
parameters is kept in order to prohibit accidental overwriting of files.