Improved saving/filing system
Instead of saving all optimisation series images and parameters for a
certain property (FOCUS for example) in one folder and numbering
duplicate images/parameters to avoid accidental overwriting, every new
optimisation series is saved in a newly created, numbered subfolder.
This looks much cleaner and it's easier to relate the images to the
according optimisation series.
However, this requires the (overloaded) saving function in
QuickAcquisitionSaver to take all the AcquisitionParameters and the
propertyValues (the values of the optimisation series) in the form of
lists.